:truck: move `internal/filterdsl` to `internal/ffi/filterdsl`
