+++ internal/ffi/filterdsl/errors.go
