+++ internal/ffi/filterdsl/ffi.go
