+++ internal/ffi/filterdsl/main.go
